pyproject.toml
[project]
name = "erp-basic"
version = "0.1.0"
description = "ERP Order Management Module"
authors = [
    { name = "Francisco Nascimento", email = "francisco.ads.dev@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.14"

dependencies = [
    "django (>=5.0,<6.0)",
    "djangorestframework (>=3.15,<4.0)",
    "mysqlclient (>=2.2,<3.0)",
    "redis (>=5.0,<6.0)",
    "django-redis (>=5.4,<6.0)",
    "drf-spectacular (>=0.27,<1.0)",
    "python-decouple (>=3.8,<4.0)",
    "cpf-cnpj-validate (>=1.4,<2.0)",
    "django-filter (>=25.2,<26.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.0,<9.0)",
    "pytest-django (>=4.8,<5.0)",
    "pytest-cov (>=5.0,<6.0)",
    "factory-boy (>=3.3,<4.0)",
    "faker (>=22.0,<25.0)",
    "black (>=24.0,<25.0)",
    "isort (>=5.13,<6.0)",
    "flake8 (>=7.0,<8.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py312"]
exclude = '''
(
  /migrations/
  | /venv/
  | /.venv/
  | /.env/
  | /.git/
  | /.idea/
  | /.vscode/
  | /docs/
)
'''

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]
skip_glob = ["*/migrations/*", "venv/*", ".venv/*", ".env/*", "docs/*"]
